+++ src/mympd_api/filesystem.c
@@ -122,19 +122,6 @@ sds mympd_api_browse_filesystem(struct t_partition_state *partition_state, sds b
 
     long entity_count = 0;
     long entities_returned = 0;
-    if (sdslen(path) > 1) {
-        sds parent_dir = sdsdup(path);
-        parent_dir = sds_dirname(parent_dir);
-
-        buffer = sdscat(buffer, "{\"Type\":\"parentDir\",\"name\":\"parentDir\",");
-        buffer = tojson_char(buffer, "uri", (parent_dir[0] == '.' ? "/" : parent_dir), false);
-        buffer = sdscatlen(buffer, "}", 1);
-        entity_count++;
-        entities_returned++;
-        real_limit++;
-        offset++;
-        FREE_SDS(parent_dir);
-    }
 
     raxIterator iter;
     raxStart(&iter, entity_list);